Splitted onoff handler into two seperate ones.
OnOff handler got the command (on|off) from the utterance and
passed in on to openhab as a command in capital letters.
This worked for englisch, but with german, there is (ein|an) for on
or even "einschalten" as a single word for "turn on".
The old approach no longer works and the new one puts the logic into
the vocal files and not into the code.